Keep aion.db replaceable while pinning wire identity
Constraint: Windows must release read-only SQLite handles immediately and native validation must remain allocation-free.
Rejected: SqliteConnection.ClearAllPools cleanup | mutates global pool state instead of preventing retention
Confidence: high
Scope-risk: moderate
Directive: Treat protocol kind as the sole wire identity; surrogate IDs and family metadata must never determine snapshot order.
Tested: 23 targeted managed tests; 7 targeted native snapshot tests; 31 full managed unit tests; 94 native tests with 2 fixture skips; VS2026 v145 Debug and Release solution builds; builder ignore and dependency audits.
Not-tested: Fixture-backed PCAP assertions because NAMTER_FIXTURE_ROOT does not contain aion2_part001.pcap.

+++ src/Namter/Namter.GameData/GameDataSnapshot.cs
@@ -10,7 +10,7 @@ public sealed record GameDataSnapshot(
     string ProtocolProfileName,
     ImmutableArray<byte> PacketMagic,
     ImmutableArray<ushort> ServerPorts,
-    FrozenDictionary<uint, ProtocolOpcode> Opcodes,
+    FrozenDictionary<ushort, ProtocolOpcode> Opcodes,
     FrozenDictionary<uint, ProtocolMessageLayout> MessageLayouts,
     FrozenDictionary<uint, Boss> Bosses,
     FrozenDictionary<uint, Dungeon> Dungeons,
